refactor: import cycle 해결
