+++ src/page/Auth/FindPassword/ChangePassword/index.tsx
@@ -2,8 +2,7 @@ import { useEffect } from 'react';
 import { useFormContext } from 'react-hook-form';
 import { useOutletContext } from 'react-router-dom';
 import { ChangePasswordForm } from 'model/auth';
-// eslint-disable-next-line
-import { OutletProps } from 'page/Auth/FindPassword/index';
+import { OutletProps } from 'page/Auth/FindPassword/entity';
 import { ReactComponent as Warning } from 'assets/svg/auth/warning.svg';
 import styles from 'page/Auth/FindPassword/index.module.scss';
